bpf: Convert whitelist_snated_egress_connections to use ct_buffer4
whitelist_snated_egress_connections contains a conntrack lookup, which
will be pulled behind a tailcall. struct ct_buffer4 is used to pass
information to the tailcall, so start using it in the mentioned
function.

A positive side effect is reusing the fields of ct_buffer allocated on
the stack of ipv4_host_policy_egress, instead of allocating new
variables on the stack of whitelist_snated_egress_connections.

While at it, also replace the error check after ct_lookup4 from IS_ERR
to ret < 0. ct_lookup4 only returns 0 or DROP_CT_CREATE_FAILED, and
having a mix of different error checks confuses the verifier in a
subsequent patch, where it can't prove that CTX_ACT_DROP is not a
possible outcome.

Signed-off-by: Maxim Mikityanskiy <maxim@isovalent.com>