bpf: Split handle_ipv6 in bpf_host.c after ct_lookup6
In order to reduce verifier complexity, split handle_ipv6 into two
functions chained by a tailcall. The first part will stop after doing
ct_lookup6 of the host firewall, and the second part will pick up the
results from a BPF map and go on.

The tailcall is saved, and the map is not used if host firewall is
disabled, so the overhead is minimal in this case.

Signed-off-by: Maxim Mikityanskiy <maxim@isovalent.com>